rss 수집건수 20건으로 증가, 오래된  뉴스 삭제 기능이 수집중에는 동작하지 않도록 설정

+++ src/main/java/devy/americano/backend/scheduler/NewsRemoveScheduler.java
@@ -1,6 +1,7 @@
 package devy.americano.backend.scheduler;
 
 import devy.americano.backend.service.NewsService;
+import devy.americano.backend.service.PublisherRssService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -12,10 +13,15 @@ import org.springframework.stereotype.Component;
 public class NewsRemoveScheduler {
     private final long FIXED_RATE_VALUE = 1 * 60 * 1000;
 
+    private final PublisherRssService publisherRssService;
     private final NewsService newsService;
 
     @Scheduled(fixedRate = FIXED_RATE_VALUE)
     public void run() {
+        if(publisherRssService.isCrawling()) {
+            return;
+        }
+
         newsService.removeOldNews();
     }
